original Skulltag 0.97c2 source from Carnevil except: - cleaned from all absolute paths and temporary files - some makefile.mgw replaced by the ones from ZDoom 2.1.7
SVN r3 (trunk)

+++ src/p_doors.cpp
@@ -34,6 +34,8 @@
 #include "a_keys.h"
 #include "i_system.h"
 #include "sc_man.h"
+#include "network.h"
+#include "sv_commands.h"
 
 IMPLEMENT_CLASS (DDoor)
 
@@ -83,16 +85,32 @@ void DDoor::Tick ()
 		// WAITING
 		if (!--m_TopCountdown)
 		{
+			// [BC] If we're the client, don't change the door's direction. It could
+			// de-sync the game. Instead, wait for the server to tell us to change
+			// the door's direction.
+			if ( NETWORK_GetState( ) == NETSTATE_CLIENT )
+				break;
+
 			switch (m_Type)
 			{
 			case doorRaise:
 				m_Direction = -1; // time to go back down
 				DoorSound (false);
+
+				// [BC] If we're the server, tell clients to change the door's direction.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+					SERVERCOMMANDS_ChangeDoorDirection( m_lDoorID, m_Direction );
+
 				break;
 				
 			case doorCloseWaitOpen:
 				m_Direction = 1;
 				DoorSound (true);
+
+				// [BC] If we're the server, tell clients to change the door's direction.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+					SERVERCOMMANDS_ChangeDoorDirection( m_lDoorID, m_Direction );
+
 				break;
 				
 			default:
@@ -105,12 +123,23 @@ void DDoor::Tick ()
 		//	INITIAL WAIT
 		if (!--m_TopCountdown)
 		{
+			// [BC] If we're the client, don't change the door's direction. It could
+			// de-sync the game. Instead, wait for the server to tell us to change
+			// the door's direction.
+			if ( NETWORK_GetState( ) == NETSTATE_CLIENT )
+				break;
+
 			switch (m_Type)
 			{
 			case doorRaiseIn5Mins:
 				m_Direction = 1;
 				m_Type = doorRaise;
 				DoorSound (true);
+			
+				// [BC] If we're the server, tell clients to change the door's direction.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+					SERVERCOMMANDS_ChangeDoorDirection( m_lDoorID, m_Direction );
+
 				break;
 				
 			default:
@@ -130,8 +159,23 @@ void DDoor::Tick ()
 				m_TopDist + m_Sector->floorplane.d));
 		}
 
+		// [BC] If we're the client, don't do any of the following. Wait for the server
+		// to tell us what to do.
+		if ( NETWORK_GetState( ) == NETSTATE_CLIENT )
+			break;
+
 		if (res == pastdest)
 		{
+			// [BC] If the sector has reached its destination, this is probably a good time to verify all the clients
+			// have the correct floor/ceiling height for this sector.
+			if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+			{
+				if ( m_Sector->floorOrCeiling == 0 )
+					SERVERCOMMANDS_SetSectorFloorPlane( m_Sector - sectors );
+				else
+					SERVERCOMMANDS_SetSectorCeilingPlane( m_Sector - sectors );
+			}
+
 			SN_StopSequence (m_Sector);
 			switch (m_Type)
 			{
@@ -139,12 +183,22 @@ void DDoor::Tick ()
 			case doorClose:
 				m_Sector->ceilingdata = NULL;	//jff 2/22/98
 				stopinterpolation (INTERP_SectorCeiling, m_Sector);
+
+				// [BC] If we're the server, tell clients to destroy the door.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+					SERVERCOMMANDS_DestroyDoor( m_lDoorID );
+
 				Destroy ();						// unlink and free
 				break;
 				
 			case doorCloseWaitOpen:
 				m_Direction = 0;
 				m_TopCountdown = m_TopWait;
+
+				// [BC] If we're the server, tell clients to change the door's direction.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+					SERVERCOMMANDS_ChangeDoorDirection( m_lDoorID, m_Direction );
+
 				break;
 				
 			default:
@@ -161,6 +215,11 @@ void DDoor::Tick ()
 			default:
 				m_Direction = 1;
 				DoorSound (true);
+
+				// [BC] If we're the server, tell clients to change the door's direction.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+					SERVERCOMMANDS_ChangeDoorDirection( m_lDoorID, m_Direction );
+
 				break;
 			}
 		}
@@ -177,20 +236,45 @@ void DDoor::Tick ()
 				m_TopDist + m_Sector->floorplane.d));
 		}
 
+		// [BC] If we're the client, don't do any of the following. Wait for the server
+		// to tell us what to do.
+		if ( NETWORK_GetState( ) == NETSTATE_CLIENT )
+			break;
+
 		if (res == pastdest)
 		{
+			// [BC] If the sector has reached its destination, this is probably a good time to verify all the clients
+			// have the correct floor/ceiling height for this sector.
+			if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+			{
+				if ( m_Sector->floorOrCeiling == 0 )
+					SERVERCOMMANDS_SetSectorFloorPlane( m_Sector - sectors );
+				else
+					SERVERCOMMANDS_SetSectorCeilingPlane( m_Sector - sectors );
+			}
+
 			SN_StopSequence (m_Sector);
 			switch (m_Type)
 			{
 			case doorRaise:
 				m_Direction = 0; // wait at top
 				m_TopCountdown = m_TopWait;
+
+				// [BC] If we're the server, tell clients to change the door's direction.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+					SERVERCOMMANDS_ChangeDoorDirection( m_lDoorID, m_Direction );
+
 				break;
 				
 			case doorCloseWaitOpen:
 			case doorOpen:
 				m_Sector->ceilingdata = NULL;	//jff 2/22/98
 				stopinterpolation (INTERP_SectorCeiling, m_Sector);
+				
+				// [BC] If we're the server, tell clients to destroy the door.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+					SERVERCOMMANDS_DestroyDoor( m_lDoorID );
+
 				Destroy ();						// unlink and free
 				break;
 				
@@ -200,6 +284,21 @@ void DDoor::Tick ()
 		}
 		break;
 	}
+}
+
+void DDoor::UpdateToClient( ULONG ulClient )
+{
+	SERVERCOMMANDS_DoDoor( m_Sector, m_Speed, m_Direction, m_LightTag, m_lDoorID, ulClient, SVCF_ONLYTHISCLIENT );
+}
+
+int DDoor::GetDirection ()
+{
+	return ( m_Direction );
+}
+
+void DDoor::SetDirection( LONG lDirection )
+{
+	m_Direction = lDirection;
 }
 
 // [RH] DoorSound: Plays door sound depending on direction and speed
@@ -295,7 +394,8 @@ DDoor::DDoor (sector_t *sector)
 //		and made them more general to support the new specials.
 
 // [RH] SpawnDoor: Helper function for EV_DoDoor
-DDoor::DDoor (sector_t *sec, EVlDoor type, fixed_t speed, int delay, int lightTag)
+// [BC] Added bNoSound. When creating doors when connecting to a server, we don't want a sound to be played.
+DDoor::DDoor (sector_t *sec, EVlDoor type, fixed_t speed, int delay, int lightTag, bool bNoSound)
 	: DMovingCeiling (sec),
   	  m_Type (type), m_Speed (speed), m_TopWait (delay), m_LightTag (lightTag)
 {
@@ -313,7 +413,9 @@ DDoor::DDoor (sector_t *sec, EVlDoor type, fixed_t speed, int delay, int lightTa
 		m_Direction = -1;
 		height = sec->FindLowestCeilingSurrounding (&spot);
 		m_TopDist = sec->ceilingplane.PointToDist (spot, height - 4*FRACUNIT);
-		DoorSound (false);
+		// [BC] Added option to create the door soundlessly.
+		if ( bNoSound == false )
+			DoorSound (false);
 		break;
 
 	case doorOpen:
@@ -321,14 +423,17 @@ DDoor::DDoor (sector_t *sec, EVlDoor type, fixed_t speed, int delay, int lightTa
 		m_Direction = 1;
 		height = sec->FindLowestCeilingSurrounding (&spot);
 		m_TopDist = sec->ceilingplane.PointToDist (spot, height - 4*FRACUNIT);
-		if (m_TopDist != sec->ceilingplane.d)
+		// [BC] Added option to create the door soundlessly.
+		if ((m_TopDist != sec->ceilingplane.d) && ( bNoSound == false ))
 			DoorSound (true);
 		break;
 
 	case doorCloseWaitOpen:
 		m_TopDist = sec->ceilingplane.d;
 		m_Direction = -1;
-		DoorSound (false);
+		// [BC] Added option to create the door soundlessly.
+		if ( bNoSound == false )
+			DoorSound (false);
 		break;
 
 	case doorRaiseIn5Mins:
@@ -351,6 +456,52 @@ DDoor::DDoor (sector_t *sec, EVlDoor type, fixed_t speed, int delay, int lightTa
 		m_BotDist = sec->ceilingplane.PointToDist (m_BotSpot, height);
 	}
 	m_OldFloorDist = sec->floorplane.d;
+
+	// [BC] Assign the door's network ID.
+	if ( NETWORK_GetState( ) != NETSTATE_CLIENT )
+		m_lDoorID = P_GetFirstFreeDoorID( );
+	else
+		m_lDoorID = -1;
+}
+
+LONG DDoor::GetID( void )
+{
+	return ( m_lDoorID );
+}
+
+void DDoor::SetID( LONG lID )
+{
+	m_lDoorID = lID;
+}
+
+DDoor::EVlDoor DDoor::GetType( void )
+{
+	return ( m_Type );
+}
+
+void DDoor::SetType( DDoor::EVlDoor Type )
+{
+	m_Type = Type;
+}
+
+LONG DDoor::GetSpeed( void )
+{
+	return ( m_Speed );
+}
+
+void DDoor::SetSpeed( LONG lSpeed )
+{
+	m_Speed = lSpeed;
+}
+
+LONG DDoor::GetLightTag( void )
+{
+	return ( m_LightTag );
+}
+
+void DDoor::SetLightTag( LONG lTag )
+{
+	m_LightTag = lTag;
 }
 
 bool EV_DoDoor (DDoor::EVlDoor type, line_t *line, AActor *thing,
@@ -359,6 +510,7 @@ bool EV_DoDoor (DDoor::EVlDoor type, line_t *line, AActor *thing,
 	bool		rtn = false;
 	int 		secnum;
 	sector_t*	sec;
+	DDoor		*pDoor;
 
 	if (lock != 0 && !P_CheckKeys (thing, lock, tag != 0))
 		return false;
@@ -372,6 +524,16 @@ bool EV_DoDoor (DDoor::EVlDoor type, line_t *line, AActor *thing,
 		if (line->sidenum[1] == NO_SIDE)			// killough
 		{
 			S_Sound (thing, CHAN_VOICE, "*usefail", 1, ATTN_NORM);
+
+			// [BC] Tell clients of the "oof" sound.
+			if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+			{
+				if ( thing && thing->player )
+					SERVERCOMMANDS_SoundActor( thing, CHAN_VOICE, "*usefail", 127, ATTN_NORM, ULONG( thing->player - players ), SVCF_SKIPTHISCLIENT );
+				else
+					SERVERCOMMANDS_SoundActor( thing, CHAN_VOICE, "*usefail", 127, ATTN_NORM );
+			}
+
 			return false;
 		}
 
@@ -393,15 +555,20 @@ bool EV_DoDoor (DDoor::EVlDoor type, line_t *line, AActor *thing,
 					{
 						door->m_Direction = 1;	// go back up
 						door->DoorSound (true);	// [RH] Make noise
+
+						// [BC] If we're the server, tell clients to change the door's direction.
+						if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+							SERVERCOMMANDS_ChangeDoorDirection( door->GetID( ), door->m_Direction );
 					}
 					else if (GET_SPAC(line->flags) != SPAC_PUSH)
 						// [RH] activate push doors don't go back down when you
 						//		run into them (otherwise opening them would be
 						//		a real pain).
 					{
-						if (!thing->player || thing->player->isbot)
+						// [BC] Added !thing, because sometimes thing can be NULL when this function
+						// is called by client_DoDoor.
+						if (!thing || !thing->player)
 							return false;	// JDC: bad guys never close doors
-											//Added by MC: Neither do bots.
 
 						door->m_Direction = -1;	// start going down immediately
 
@@ -412,6 +579,11 @@ bool EV_DoDoor (DDoor::EVlDoor type, line_t *line, AActor *thing,
 						{
 							door->DoorSound (false);
 						}
+
+						// [BC] If we're the server, tell clients to change the door's direction.
+						if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+							SERVERCOMMANDS_ChangeDoorDirection( door->GetID( ), door->m_Direction );
+
 						return true;
 					}
 					else
@@ -422,8 +594,13 @@ bool EV_DoDoor (DDoor::EVlDoor type, line_t *line, AActor *thing,
 			}
 			return false;
 		}
-		if (new DDoor (sec, type, speed, delay, lightTag))
+		if ( pDoor = new DDoor (sec, type, speed, delay, lightTag))
+		{
+			if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+				SERVERCOMMANDS_DoDoor( sec, speed, pDoor->GetDirection( ), lightTag, pDoor->GetID( ));
+
 			rtn = true;
+		}
 	}
 	else
 	{	// [RH] Remote door
@@ -436,8 +613,13 @@ bool EV_DoDoor (DDoor::EVlDoor type, line_t *line, AActor *thing,
 			if (sec->ceilingdata)
 				continue;
 
-			if (new DDoor (sec, type, speed, delay, lightTag))
+			if ( pDoor = new DDoor (sec, type, speed, delay, lightTag))
+			{
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+					SERVERCOMMANDS_DoDoor( sec, speed, pDoor->GetDirection( ), lightTag, pDoor->GetID( ));
+
 				rtn = true;
+			}
 		}
 				
 	}
@@ -474,6 +656,52 @@ void P_SpawnDoorRaiseIn5Mins (sector_t *sec)
 {
 	sec->special = 0;
 	new DDoor (sec, DDoor::doorRaiseIn5Mins, 2*FRACUNIT, TICRATE*30/7, 0);
+}
+
+//*****************************************************************************
+//
+DDoor *P_GetDoorByID( LONG lID )
+{
+	DDoor	*pDoor;
+
+	TThinkerIterator<DDoor>		Iterator;
+
+	while (( pDoor = Iterator.Next( )))
+	{
+		if ( pDoor->GetID( ) == lID )
+			return ( pDoor );
+	}
+
+	return ( NULL );
+}
+
+//*****************************************************************************
+//
+LONG P_GetFirstFreeDoorID( void )
+{
+	LONG		lIdx;
+	DDoor		*pDoor;
+	bool		bIDIsAvailable;
+
+	for ( lIdx = 0; lIdx < 8192; lIdx++ )
+	{
+		TThinkerIterator<DDoor>		Iterator;
+
+		bIDIsAvailable = true;
+		while (( pDoor = Iterator.Next( )))
+		{
+			if ( pDoor->GetID( ) == lIdx )
+			{
+				bIDIsAvailable = false;
+				break;
+			}
+		}
+
+		if ( bIDIsAvailable )
+			return ( lIdx );
+	}
+
+	return ( -1 );
 }
 
 // Strife's animated doors. Based on Doom's unused sliding doors, but slightly different.
@@ -547,6 +775,17 @@ bool DAnimatedDoor::StartClosing ()
 
 	m_Line1->flags |= ML_BLOCKING;
 	m_Line2->flags |= ML_BLOCKING;
+
+	// [BC] If we're the server, tell clients to alter this line's blocking status.
+	if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+	{
+		SERVERCOMMANDS_SetLineBlocking( ULONG( m_Line1 - lines ));
+		SERVERCOMMANDS_SetLineBlocking( ULONG( m_Line2 - lines ));
+
+		// Also, tell clients to move the ceiling.
+		SERVERCOMMANDS_SetSectorCeilingPlane( ULONG( m_Sector - sectors ));
+	}
+
 	if (ani.CloseSound != NULL)
 	{
 		SN_StartSequence (m_Sector, ani.CloseSound, 1);
@@ -577,6 +816,13 @@ void DAnimatedDoor::Tick ()
 				m_Line1->flags &= ~ML_BLOCKING;
 				m_Line2->flags &= ~ML_BLOCKING;
 
+				// [BC] If we're the server, tell clients to alter this line's blocking status.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+				{
+					SERVERCOMMANDS_SetLineBlocking( ULONG( m_Line1 - lines ));
+					SERVERCOMMANDS_SetLineBlocking( ULONG( m_Line2 - lines ));
+				}
+
 				if (m_Delay == 0)
 				{
 					m_Sector->ceilingdata = NULL;
@@ -597,6 +843,18 @@ void DAnimatedDoor::Tick ()
 				sides[m_Line2->sidenum[0]].midtexture =
 				sides[m_Line2->sidenum[1]].midtexture =
 					ani.TextureFrames[m_Frame];
+
+				// [BC] Mark this line's textures as having been changed.
+				m_Line1->ulTexChangeFlags |= TEXCHANGE_FRONTMEDIUM|TEXCHANGE_BACKMEDIUM;
+				m_Line2->ulTexChangeFlags |= TEXCHANGE_FRONTMEDIUM|TEXCHANGE_BACKMEDIUM;
+
+				// [BC] If we're the server, tell clients that these lines' texture
+				// has changed.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+				{
+					SERVERCOMMANDS_SetLineTexture( ULONG( m_Line1 - lines ));
+					SERVERCOMMANDS_SetLineTexture( ULONG( m_Line2 - lines ));
+				}
 			}
 		}
 		break;
@@ -619,6 +877,11 @@ void DAnimatedDoor::Tick ()
 			{
 				// IF DOOR IS DONE CLOSING...
 				MoveCeiling (2048*FRACUNIT, m_BotDist, -1);
+
+				// [BC] If we're the server, tell clients to move the ceiling.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+					SERVERCOMMANDS_SetSectorCeilingPlane( ULONG( m_Sector - sectors ));
+
 				m_Sector->ceilingdata = NULL;
 				Destroy ();
 				break;
@@ -633,6 +896,18 @@ void DAnimatedDoor::Tick ()
 				sides[m_Line2->sidenum[0]].midtexture =
 				sides[m_Line2->sidenum[1]].midtexture =
 					ani.TextureFrames[m_Frame];
+
+				// [BC] Mark this line's textures as having been changed.
+				m_Line1->ulTexChangeFlags |= TEXCHANGE_FRONTMEDIUM|TEXCHANGE_BACKMEDIUM;
+				m_Line2->ulTexChangeFlags |= TEXCHANGE_FRONTMEDIUM|TEXCHANGE_BACKMEDIUM;
+
+				// [BC] If we're the server, tell clients that these lines' texture
+				// has changed.
+				if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+				{
+					SERVERCOMMANDS_SetLineTexture( ULONG( m_Line1 - lines ));
+					SERVERCOMMANDS_SetLineTexture( ULONG( m_Line2 - lines ));
+				}
 			}
 		}
 		break;
@@ -712,6 +987,18 @@ DAnimatedDoor::DAnimatedDoor (sector_t *sec, line_t *line, int speed, int delay)
 	sides[m_Line1->sidenum[0]].midtexture = picnum;
 	sides[m_Line2->sidenum[0]].midtexture = picnum;
 
+	// [BC] Mark this line's textures as having been changed.
+	m_Line1->ulTexChangeFlags |= TEXCHANGE_FRONTMEDIUM|TEXCHANGE_BACKMEDIUM;
+	m_Line2->ulTexChangeFlags |= TEXCHANGE_FRONTMEDIUM|TEXCHANGE_BACKMEDIUM;
+
+	// [BC] If we're the server, tell clients that these lines' texture
+	// has changed.
+	if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+	{
+		SERVERCOMMANDS_SetLineTexture( ULONG( m_Line1 - lines ));
+		SERVERCOMMANDS_SetLineTexture( ULONG( m_Line2 - lines ));
+	}
+
 	// don't forget texture scaling here!
 	FTexture *tex = TexMan[picnum];
 	topdist = tex ? (tex->GetHeight() * 8 / (tex->ScaleY ? tex->ScaleY : 8)) : 64;	
@@ -726,12 +1013,25 @@ DAnimatedDoor::DAnimatedDoor (sector_t *sec, line_t *line, int speed, int delay)
 	m_Frame = 0;
 	m_Line1->flags |= ML_BLOCKING;
 	m_Line2->flags |= ML_BLOCKING;
+
+	// [BC] If we're the server, tell clients that these lines' blocking status
+	// has changed.
+	if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+	{
+		SERVERCOMMANDS_SetLineBlocking( ULONG( m_Line1 - lines ));
+		SERVERCOMMANDS_SetLineBlocking( ULONG( m_Line2 - lines ));
+	}
+
 	m_BotDist = m_Sector->ceilingplane.d;
 	MoveCeiling (2048*FRACUNIT, topdist, 1);
 	if (DoorAnimations[m_WhichDoorIndex].OpenSound != NULL)
 	{
 		SN_StartSequence (m_Sector, DoorAnimations[m_WhichDoorIndex].OpenSound, 1);
 	}
+
+	// [BC] If we're the server, tell clients to move the ceiling.
+	if ( NETWORK_GetState( ) == NETSTATE_SERVER )
+		SERVERCOMMANDS_SetSectorCeilingPlane( ULONG( m_Sector - sectors ));
 }
 
 //============================================================================
